Fixed ai model
